fix(integer): clear -Wsign-conversion in integer/primes/lns (#1283) (#1284)
* fix(integer): clear -Wsign-conversion in integer/primes/lns (#1283)

Sub-issue of Epic #1279 (-Wsign-conversion cleanup). All fixes are
value-preserving; each casts a value already proven non-negative /
in-range at the site, or replaces a signedness-changing literal.

integer_impl.hpp (shift operators <<=/>>=):
- guard the block-shift index arithmetic where int bitsToShift/blockShift
  mixes with unsigned bitsInBlock/nbits/MSU; bitsToShift is > 0 past the
  guards, so static_cast<unsigned> is exact
- upper-block mask now uses the shared bit_high_mask<bt>(#1261) helper,
  which also states the truncation explicitly (was a raw 64-bit literal
  shift that narrowed implicitly)
- assign(int64_t): low-bit test v & 0x1ull -> (v & 1) != 0; masking with a
  signed 1 keeps the operation in int64 (correct for negative v) instead
  of promoting v to uint64
- operator<<: width padding casts std::streamsize width (guarded
  width > s.size(), so positive) before subtracting the string size

primes.hpp: printPrimes column counter int -> size_t (compared against
size_t COL_WIDTH/PAGE_WIDTH).

lns_impl.hpp: setbit(i + shiftLeft) with shiftLeft used as a left-shift
count (>= 0 on this path) -> cast to unsigned.

Verified -Wsign-conversion clean for all three headers on gcc and clang;
no new -Wconversion introduced. integer shift_left/shift_right/division/
remainder and lns conversion regressions PASS on both compilers.

Relates to #1279


* fix(integer): address CodeRabbit review on #1283

- revert the unsigned-path low-bit test (convert_unsigned) to its original
  `v & 0x1ull`: `v` is uint64_t there, so it never triggered
  -Wsign-conversion; a global sed had changed it and attached a comment
  that (correctly, per review) described the signed path
- keep only the signed-path change (int64_t v) where the comment is accurate
- wrap two trailing comments that pushed the shifted-mask lines past the
  120-column limit onto their own lines
- hoist the output-padding count into a `pad` local so the append/insert
  lines stay within the column limit

Relates to #1279
Resolves #1283


---------

+++ include/sw/universal/number/integer/primes.hpp
@@ -173,7 +173,7 @@ void printPrimes(const std::vector< IntegerType >& v) {
 		number /= 10;
 	}
 	std::cout << "largest prime: " << v[nrPrimes - 1] << " is " << COL_WIDTH - 1 << " decades\n";
-	int column = 1;
+	size_t column = 1;
 	for (auto p : v) {
 		std::cout << std::setw(COL_WIDTH) << p;
 		if (column * COL_WIDTH < PAGE_WIDTH) {
